# refactor(ivy): remove styling state storage and introduce direct style writing