--- a/packages/core/src/debug/debug_node.ts
+++ b/packages/core/src/debug/debug_node.ts
@@ -12,10 +12,9 @@ import {CONTAINER_HEADER_OFFSET, LContainer, NATIVE} from '../render3/interfaces
 import {TElementNode, TNode, TNodeFlags, TNodeType} from '../render3/interfaces/node';
 import {isComponentHost, isLContainer} from '../render3/interfaces/type_checks';
 import {LView, PARENT, TData, TVIEW, T_HOST} from '../render3/interfaces/view';
-import {TStylingContext} from '../render3/styling_next/interfaces';
-import {stylingMapToStringMap} from '../render3/styling_next/map_based_bindings';
+import {StylingMapArray, TStylingContext} from '../render3/styling_next/interfaces';
 import {NodeStylingDebug} from '../render3/styling_next/styling_debug';
-import {isStylingContext} from '../render3/styling_next/util';
+import {isStylingContext, stylingMapToStringMap} from '../render3/styling_next/util';
 import {getComponent, getContext, getInjectionTokens, getInjector, getListeners, getLocalRefs, isBrowserEvents, loadLContext} from '../render3/util/discovery_utils';
 import {INTERPOLATION_DELIMITER, renderStringify} from '../render3/util/misc_utils';
 import {findComponentView} from '../render3/util/view_traversal_utils';
@@ -431,11 +430,11 @@ function _getStylingDebugInfo(element: any, isClassBased: boolean) {
   if (isClassBased) {
     return isStylingContext(tNode.classes) ?
         new NodeStylingDebug(tNode.classes as TStylingContext, lView, true).values :
-        stylingMapToStringMap(tNode.classes);
+        stylingMapToStringMap(tNode.classes as StylingMapArray | null);
   } else {
     return isStylingContext(tNode.styles) ?
         new NodeStylingDebug(tNode.styles as TStylingContext, lView, false).values :
-        stylingMapToStringMap(tNode.styles);
+        stylingMapToStringMap(tNode.styles as StylingMapArray | null);
   }
 }