--- a/packages/core/src/render3/features/inherit_definition_feature.ts
+++ b/packages/core/src/render3/features/inherit_definition_feature.ts
@@ -11,7 +11,6 @@ import {fillProperties} from '../../util/property';
 import {EMPTY_ARRAY, EMPTY_OBJ} from '../empty';
 import {ComponentDef, ContentQueriesFunction, DirectiveDef, DirectiveDefFeature, HostBindingsFunction, RenderFlags, ViewQueriesFunction} from '../interfaces/definition';
 import {isComponentDef} from '../interfaces/type_checks';
-import {adjustActiveDirectiveSuperClassDepthPosition} from '../state';
 
 import {ɵɵNgOnChangesFeature} from './ng_onchanges_feature';
 
@@ -177,24 +176,8 @@ function inheritHostBindings(
   // to ensure we don't inherit it twice.
   if (superHostBindings !== prevHostBindings) {
     if (prevHostBindings) {
-      // because inheritance is unknown during compile time, the runtime code
-      // needs to be informed of the super-class depth so that instruction code
-      // can distinguish one host bindings function from another. The reason why
-      // relying on the directive uniqueId exclusively is not enough is because the
-      // uniqueId value and the directive instance stay the same between hostBindings
-      // calls throughout the directive inheritance chain. This means that without
-      // a super-class depth value, there is no way to know whether a parent or
-      // sub-class host bindings function is currently being executed.
       definition.hostBindings = (rf: RenderFlags, ctx: any, elementIndex: number) => {
-        // The reason why we increment first and then decrement is so that parent
-        // hostBindings calls have a higher id value compared to sub-class hostBindings
-        // calls (this way the leaf directive is always at a super-class depth of 0).
-        adjustActiveDirectiveSuperClassDepthPosition(1);
-        try {
-          superHostBindings(rf, ctx, elementIndex);
-        } finally {
-          adjustActiveDirectiveSuperClassDepthPosition(-1);
-        }
+        superHostBindings(rf, ctx, elementIndex);
         prevHostBindings(rf, ctx, elementIndex);
       };
     } else {