--- a/packages/core/src/render3/index.ts
+++ b/packages/core/src/render3/index.ts
@@ -114,8 +114,6 @@ export {
   ɵɵstylePropInterpolateV,
 
   ɵɵstyleSanitizer,
-  ɵɵstyling,
-  ɵɵstylingApply,
   ɵɵtemplate,
 
   ɵɵtext,