--- a/packages/core/src/render3/instructions/element.ts
+++ b/packages/core/src/render3/instructions/element.ts
@@ -138,11 +138,11 @@ export function ɵɵelementEnd(): void {
     }
   }
 
-  if (hasClassInput(tNode) && tNode.classes) {
+  if (hasClassInput(tNode)) {
     setDirectiveStylingInput(tNode.classes, lView, tNode.inputs !['class']);
   }
 
-  if (hasStyleInput(tNode) && tNode.styles) {
+  if (hasStyleInput(tNode)) {
     setDirectiveStylingInput(tNode.styles, lView, tNode.inputs !['style']);
   }
 }
@@ -236,11 +236,12 @@ export function ɵɵelementHostAttrs(attrs: TAttributes) {
 }
 
 function setDirectiveStylingInput(
-    context: TStylingContext | StylingMapArray, lView: LView, stylingInputs: (string | number)[]) {
+    context: TStylingContext | StylingMapArray | null, lView: LView,
+    stylingInputs: (string | number)[]) {
   // older versions of Angular treat the input as `null` in the
   // event that the value does not exist at all. For this reason
   // we can't have a styling value be an empty string.
-  const value = getInitialStylingValue(context) || null;
+  const value = (context && getInitialStylingValue(context)) || null;
 
   // Ivy does an extra `[class]` write with a falsy value since the value
   // is applied during creation mode. This is a deviation from VE and should