--- a/packages/core/src/render3/styling_next/map_based_bindings.ts
+++ b/packages/core/src/render3/styling_next/map_based_bindings.ts
@@ -225,7 +225,7 @@ function innerSyncStylingMap(
             } else {
               finalValue = sanitizer ?
                   sanitizer(prop, value, StyleSanitizeMode.ValidateAndSanitize) :
-                  unwrapSafeValue(value);
+                  (value ? unwrapSafeValue(value) : null);
             }
 
             applyStylingFn(renderer, element, prop, finalValue, bindingIndexToApply);