--- a/packages/core/src/render3/styling_next/state.ts
+++ b/packages/core/src/render3/styling_next/state.ts
@@ -5,11 +5,11 @@
 * Use of this source code is governed by an MIT-style license that can be
 * found in the LICENSE file at https://angular.io/license
 */
+import {RElement} from '../interfaces/renderer';
+import {TEMPLATE_DIRECTIVE_INDEX} from './util';
 
 /**
  * --------
- *
- * // TODO(matsko): add updateMask info
  *
  * This file contains all state-based logic for styling in Angular.
  *
@@ -23,79 +23,92 @@
  * exited in change detection (once all the instructions are run for
  * that element).
  *
- * There are, however, situations where the state-based values
- * need to be stored and used at a later point. This ONLY occurs when
- * there are template-level as well as host-binding-level styling
- * instructions on the same element. The example below shows exactly
- * what could be:
- *
- * ```html
- * <!-- two sources of styling: the template and the directive -->
- * <div [style.width]="width" dir-that-sets-height></div>
- * ```
- *
- * If and when this situation occurs, the current styling state is
- * stored in a storage map value and then later accessed once the
- * host bindings are evaluated. Once styling for the current element
- * is over then the map entry will be cleared.
- *
  * To learn more about the algorithm see `TStylingContext`.
  *
  * --------
  */
 
-let _stylingState: StylingState|null = null;
-const _stateStorage = new Map<any, StylingState>();
-
-// this value is not used outside this file and is only here
-// as a caching check for when the element changes.
-let _stylingElement: any = null;
-
 /**
  * Used as a state reference for update values between style/class binding instructions.
+ *
+ * In addition to storing the element and bit-mask related values, the state also
+ * stores the `sourceIndex` value. The `sourceIndex` value is an incremented value
+ * that identifies what "source" (i.e. the template, a specific directive by index or
+ * component) is currently applying its styling bindings to the element.
  */
 export interface StylingState {
+  /** The element that is currently being processed */
+  element: RElement|null;
+
+  /** The directive index that is currently active (`0` === template) */
+  directiveIndex: number;
+
+  /** The source (column) index that is currently active (`0` === template) */
+  sourceIndex: number;
+
+  /** The classes update bit mask value that is processed during each class binding */
   classesBitMask: number;
+
+  /** The classes update bit index value that is processed during each class binding */
   classesIndex: number;
+
+  /** The styles update bit mask value that is processed during each style binding */
   stylesBitMask: number;
+
+  /** The styles update bit index value that is processed during each style binding */
   stylesIndex: number;
 }
 
-export const STYLING_INDEX_START_VALUE = 1;
-export const BIT_MASK_START_VALUE = 0;
+// these values will get filled in the very first time this is accessed...
+const _state: StylingState = {
+  element: null,
+  directiveIndex: -1,
+  sourceIndex: -1,
+  classesBitMask: -1,
+  classesIndex: -1,
+  stylesBitMask: -1,
+  stylesIndex: -1,
+};
 
-export function getStylingState(element: any, readFromMap?: boolean): StylingState {
-  if (!_stylingElement || element !== _stylingElement) {
-    _stylingElement = element;
-    if (readFromMap) {
-      _stylingState = _stateStorage.get(element) || null;
-      ngDevMode && ngDevMode.stylingReadPersistedState++;
-    }
-    _stylingState = _stylingState || {
-      classesBitMask: BIT_MASK_START_VALUE,
-      classesIndex: STYLING_INDEX_START_VALUE,
-      stylesBitMask: BIT_MASK_START_VALUE,
-      stylesIndex: STYLING_INDEX_START_VALUE,
-    };
+const BIT_MASK_START_VALUE = 0;
+
+// the `0` start value is reserved for [map]-based entries
+const INDEX_START_VALUE = 1;
+
+/**
+ * Returns (or instantiates) the styling state for the given element.
+ *
+ * Styling state is accessed and processed each time a style or class binding
+ * is evaluated.
+ *
+ * If and when the provided `element` doesn't match the current element in the
+ * state then this means that styling was recently cleared or the element has
+ * changed in change detection. In both cases the styling state is fully reset.
+ *
+ * If and when the provided `directiveIndex` doesn't match the current directive
+ * index in the state then this means that a new source has introduced itself into
+ * the styling code (or, in other words, another directive or component has started
+ * to apply its styling host bindings to the element).
+ */
+export function getStylingState(element: RElement, directiveIndex: number): StylingState {
+  if (_state.element !== element) {
+    _state.element = element;
+    _state.directiveIndex = directiveIndex;
+    _state.sourceIndex = directiveIndex === TEMPLATE_DIRECTIVE_INDEX ? 0 : 1;
+    _state.classesBitMask = BIT_MASK_START_VALUE;
+    _state.classesIndex = INDEX_START_VALUE;
+    _state.stylesBitMask = BIT_MASK_START_VALUE;
+    _state.stylesIndex = INDEX_START_VALUE;
+  } else if (_state.directiveIndex !== directiveIndex) {
+    _state.directiveIndex = directiveIndex;
+    _state.sourceIndex++;
   }
-  return _stylingState !;
+  return _state;
 }
 
+/**
+ * Clears the styling state so that it can be used by another element's styling code.
+ */
 export function resetStylingState() {
-  _stylingState = null;
-  _stylingElement = null;
-}
-
-export function storeStylingState(element: any, state: StylingState) {
-  ngDevMode && ngDevMode.stylingWritePersistedState++;
-  _stateStorage.set(element, state);
-}
-
-export function deleteStylingStateFromStorage(element: any) {
-  _stateStorage.delete(element);
-}
-
-export function resetAllStylingState() {
-  resetStylingState();
-  _stateStorage.clear();
+  _state.element = null;
 }