… aio/scripts/_payload-limits.json
… modules/benchmarks/src/largetable/ng2/table.ts
… modules/benchmarks/src/tree/render3_function/index.ts
… packages/compiler-cli/test/compliance/r3_compiler_compliance_spec.ts
… packages/compiler-cli/test/compliance/r3_view_compiler_styling_spec.ts
… packages/compiler-cli/test/ngtsc/ngtsc_spec.ts
… packages/core/src/render3/jit/environment.ts
… packages/core/src/render3/state.ts
… packages/core/src/render3/styling_next/bindings.ts
… packages/core/src/render3/styling_next/instructions.ts
… packages/core/src/render3/styling_next/interfaces.ts
… packages/core/src/render3/styling_next/styling_debug.ts
… packages/core/src/render3/styling_next/util.ts
… packages/core/src/util/ng_dev_mode.ts
… packages/core/test/acceptance/styling_next_spec.ts
… packages/core/test/acceptance/styling_spec.ts
… packages/core/test/bundling/cyclic_import/bundle.golden_symbols.json
… packages/core/test/bundling/hello_world/bundle.golden_symbols.json
… packages/core/test/bundling/todo/bundle.golden_symbols.json
… packages/core/test/render3/instructions_spec.ts
… packages/core/test/render3/integration_spec.ts
… packages/core/test/render3/perf/map_based_style_and_class_bindings/index.ts
… packages/core/test/render3/perf/style_and_class_bindings/index.ts
… packages/core/test/render3/perf/style_binding/index.ts
… packages/core/test/render3/styling_next/map_based_bindings_spec.ts
… packages/core/test/render3/styling_next/styling_context_spec.ts
… packages/core/test/render3/styling_next/styling_debug_spec.ts
… tools/public_api_guard/core/core.d.ts