This patch is a final major refactor in styling Angular.

This PR includes three main fixes:
- All temporary state taht is persisted between template style/class application
  and style/class application in host bindings is now removed.
- Removes the `styling()` and `stylingApply()` instructions.
- Introduces a "direct apply" mode that is used apply prop-based
  style/class in the event that there are no map-based bindings as
  well as property collisions.