@matsko 2019-08-29
Current Benchmarks:

```
[VE]
CREATE = time: 41.33235099332234 ms / iteration
CREATE_DIFF = 0%
UPDATE = time: 6.8551655627948245 ms / iteration
UPDATE_DIFF = 0%

[IVY - Master]
CREATE = time: 53.1655298012547 ms / iteration
CREATE_DIFF = 28%
UPDATE = time: 7.157582781371654 ms / iteration
UPDATE_DIFF = 4%

[IVY - THIS PR]
CREATE = time: 45.67341059578362 ms / iteration
CREATE_DIFF = 10%
UPDATE = time: 7.0579470197815874 ms / iteration
UPDATE_DIFF = 2%
```