@matsko 2019-08-30
# Benchmarks

## Large Table

### VE
CREATE = 38.435331126270256 ms
UPDATE = 6.078046357986943 ms

### IVY (based on master)
CREATE = 54.54857615973347 ms (40% slower)
UPDATE = 7.8921523179345465 ms (16% slower)

### IVY (based on this PR)
CREATE = 46.36072847700758 ms (20% slower)
UPDATE = 7.016092714709415 ms (16% slower)

## Micro Benchmarks

### [style.prop] (result = 45% faster)
MASTER = 20.81s user 0.21s system 103% cpu 20.339 total
THIS PR = 11.33s user 0.08s system 100% cpu 11.372 total

### [style.prop] and [class.prop] (result = 28% faster)
MASTER = 30.63s user 0.27s system 102% cpu 30.244 total
THIS PR = 21.80s user 0.11s system 100% cpu 21.882 total

### [style] and [class] (result = same)
MASTER = 68.91s user 1.22s system 104% cpu 1:07.15 total
THIS PR = 67.72s user 1.07s system 102% cpu 1:06.83 total